Fix response parsing for non ascii strings
The response was being read assuming that the number of characters is
equal to the number of byets. The problem is fixed.

Now (response) gets a DataInput instead of a reader. Reader is not the
right interface for (response) since it's used to read chars. We need to
read bytes, because Redis protocol uses numbers of bytes not characters.

We still need a fix for generating non ascii strings in queries.