Handle socket failure reasonably well
If a socket throws an exception reading or writing, close it and open
a new one in the agent, but do not retry the query.

Throwables thrown in the agent will be manually rethrown in the caller
thread.

I'm not entirely crazy about how this worked out. Some pieces I'm not
sure are right include:

- should we really be catching Throwable? should that be Exception? if
  so, is it a problem that the promise will never be delivered?
- is manually checking the type of the promise result insanely bad?
  maybe use a protocol there to make it fast/non-reflective
- should we include an error handler in agents that logs? is it ok to
  use clojure.tools.logging? seems like it'd be useful.

Also uncomment the timeout test - working locally, may need some love
for travis-ci